requirements/base.txt


# Bleeding edge Django
django==1.10.2

# Configuration
django-environ==0.4.0
whitenoise==3.2.2


# Forms
django-braces==1.9.0
django-crispy-forms==1.6.0

# Models
django-model-utils==2.6

# Images
Pillow==3.4.1

# For user registration
django-allauth==0.27.0


# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2016.7

# Redis support
django-redis==4.5.0
redis>=2.10.5


# Custom requirements
python-dateutil==2.5.3
six==1.10.0
mysqlclient==1.3.9
django-filter==0.15.3
django-tinymce==2.4.0
djangorestframework==3.5.3
git://github.com/kavdev/dj-stripe.git@1.0.0#egg=dj-stripe
mock-django==0.6.10
